requirements.txt
pymongo==3.11.4
google-cloud-bigquery==2.16.1